Extracted MaterialData into its own script
