+++ Sol/Sol.vcxproj
@@ -165,6 +165,7 @@ IF EXIST ..\Sol\thirdparty\assimp\lib\*.dll\ (xcopy /Q /E /Y /I ..\Sol\thirdpart
     <ClInclude Include="Source\GalaxyDraw\Interfaces\Buffer.h" />
     <ClInclude Include="Source\GalaxyDraw\Interfaces\Framebuffer.h" />
     <ClInclude Include="Source\GalaxyDraw\Interfaces\GalaxyMacros.h" />
+    <ClInclude Include="Source\GalaxyDraw\Interfaces\MaterialData.h" />
     <ClInclude Include="Source\GalaxyDraw\Interfaces\MeshRenderData.h" />
     <ClInclude Include="Source\GalaxyDraw\Interfaces\IMaterial.h" />
     <ClInclude Include="Source\GalaxyDraw\Interfaces\InstanceData.h" />
