+++ Sol/Source/GalaxyDraw/Interfaces/MaterialData.h
@@ -1,0 +1,20 @@
+#pragma once
+
+
+#include "MeshRenderData.h"
+#include "Shader.h"
+
+#include "Sol/SolDefines.h"
+#include "Sol/Core/KeyedVector.h"
+
+namespace GalaxyDraw
+{
+	struct MaterialData
+	{
+		std::string Name;
+		KeyedVector<std::string, MeshRenderData> MeshDataCollections;
+		std::vector<EntityID> EntitiesUsingMat;
+		std::string DiffuseTexturePath;
+		std::shared_ptr<Shader> Shader;
+	};
+}
